modified

+++ src/SYSTEM_DESIGN/OOPS/WrapperExample.java
@@ -4,7 +4,7 @@ public class WrapperExample {
 
     public static void main(String[] args) {
         Integer a = 20;
-        System.out.println(a.longValue());
+        System.out.println(a.byteValue());
     }
 
 }
